Rework Language Containers of the Localisation System
Dictionary approach is not actually needed because there are 7160 officially known living languages worldwide, and even if all of them were in this program, it won't be have much of an performance impact that our Dictionaries gave over the new List.

Also made some QoL improvements are mainly getting to code to fit new C#/.NET guidelines.